+++ packages/moonstone/Panels/Viewport.js
@@ -91,15 +91,6 @@ const ViewportBase = class extends React.Component {
 		return true;
 	}
 
-	mayBlurSpotlight = () => {
-		const current = Spotlight.getCurrent();
-		if (current && !this.props.noAnimation && this.node.contains(current)) {
-			current.blur();
-		}
-
-		return true;
-	}
-
 	handle = handle.bind(this)
 
 	handleTransition = this.handle(
@@ -111,7 +102,6 @@ const ViewportBase = class extends React.Component {
 	handleWillTransition = this.handle(
 		forward('onWillTransition'),
 		this.addTransitioningClass,
-		this.mayBlurSpotlight,
 		Spotlight.pause
 	)
 
@@ -143,6 +133,7 @@ const ViewportBase = class extends React.Component {
 				className={className}
 				component="main"
 				duration={250}
+				enteringDelay={100} // TODO: Can we remove this?
 				enteringProp={enteringProp}
 				index={index}
 				noAnimation={noAnimation}